Move settings out of dashboard folder
