+++ packages/desktop/views/dashboard/index.ts
@@ -1,6 +1,5 @@
 export * from './collectibles'
 export * from './developer'
-export * from './settings'
 export * from './old-wallet'
 export * from './governance'
 export * from './send-flow'
