+++ packages/desktop/views/index.ts
@@ -2,4 +2,5 @@ export { default as Splash } from './Splash.svelte'
 
 export * from './dashboard'
 export * from './login'
+export * from './settings'
 export * from './update-stronghold'
