+++ packages/desktop/views/settings/views/SettingsListForCategory.svelte
