+++ packages/desktop/views/settings/views/SettingsSidebar.svelte
