+++ packages/desktop/views/settings/views/advanced/HiddenAccounts.svelte
