+++ packages/desktop/views/settings/views/collectibles/MaxMediaDownloadTime.svelte
