+++ packages/desktop/views/settings/views/collectibles/MaxMediaSize.svelte
