+++ packages/desktop/views/settings/views/collectibles/collectibles-settings.constant.js
