+++ packages/desktop/views/settings/views/general/CrashReporting.svelte
