+++ packages/desktop/views/settings/views/general/DeepLinks.svelte
