+++ packages/desktop/views/settings/views/general/Notifications.svelte
