+++ packages/desktop/views/settings/views/general/general-settings.constant.ts
