+++ packages/desktop/views/settings/views/help/Diagnostics.svelte
