+++ packages/desktop/views/settings/views/help/ErrorLog.svelte
