+++ packages/desktop/views/settings/views/help/HelpSection.svelte
