+++ packages/desktop/views/settings/views/help/help-settings.constant.js
