+++ packages/desktop/views/settings/views/profile/Currency.svelte
