+++ packages/desktop/views/settings/views/profile/DeleteProfile.svelte
