+++ packages/desktop/views/settings/views/security/ExportStronghold.svelte
