+++ packages/desktop/views/settings/views/security/StrongholdTimeout.svelte
